common,jinja2util: move common.j2_render() to new "jinja2util.py" module
"common" should not depend on additional pip modules. We should be able
to copy it somewhere and use with minimal effort (e.g. not need to
install additional pip modules).

This is almost the case, except for common.j2_render() and
common.j2_render_data(). Move those two functions to a new module
"jinja2util.py".
